Support Kafka client 4.2.0 while retaining 4.1.1 support.
